fix(v3.3.1): add ruff config + restore _RAG_DISABLED re-export
- [tool.ruff] config in pyproject.toml: ignore pre-existing rules
  (BLE001, DTZ005, RUF012, S110, etc) that ruff 0.16 enables by default
- Restore _RAG_DISABLED re-export (ruff --fix removed it as 'unused')
- ruff clean, pyright 0 errors, 2622 tests passing

+++ conscio/agency/skills.py
@@ -196,9 +196,9 @@ class SkillLibrary:
         exemplar teaches tool/arg choice, not response syntax)."""
         steps = json.loads(row["plan_template"])
         total = row["successes"] + row["failures"]
-        rate = int(round(100 * row["successes"] / total)) if total else 100
-        lines = [f"Past successful plan for a similar goal "
-                 f"(success rate {rate}%, used {row['uses']}x):"]
+        rate = round(100 * row["successes"] / total) if total else 100
+        lines = [(f"Past successful plan for a similar goal "
+                 f"(success rate {rate}%, used {row['uses']}x):")]
         for step in steps:
             if tier == "T3":
                 lines.append(f"TOOL: {step['tool']}")
